Fix foreign value imports
Leaving this as a note to myself for later:

It appears that the issue wasn't with the imports themselves, but with
the way that I was using them. This might just be a type-related thing
that I don't understand, but a case statment doesn't work for
equality-checking these values (and I would assume CUInts in general),
so `(==)` must be used.